requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.2.2
python-decouple==3.6
